+++ src/main/java/org/sonarsource/plugins/rtechnicaldebt/measures/cumulative/ComputeAverageCE.java
@@ -26,14 +26,14 @@ public class ComputeAverageCE implements MeasureComputer {
     // measure is already defined on files by {@link SetSizeOnFilesSensor}
     // in scanner stack
     if (context.getComponent().getType() != Component.Type.FILE) {
-      int sum = 0;
+      double sum = 0;
       int count = 0;
       for (Measure child : context.getChildrenMeasures(EFFERENT_COUPLING.key())) {
         sum += child.getIntValue();
         count++;
       }
-      int average = count == 0 ? 0 : sum / count;
-      context.addMeasure(EFFERENT_COUPLING.key(), average);
+      double average = count == 0 ? 0 : sum / count;
+      context.addMeasure(EFFERENT_COUPLING.key(),(float) average);
     }
   }
 }
